--- a/usaspending_api/awards/models_matviews.py
+++ b/usaspending_api/awards/models_matviews.py
@@ -354,6 +354,49 @@ class SummaryTransactionMonthView(models.Model):
         db_table = 'summary_transaction_month_view'
 
 
+class SummaryTransactionGeoView(models.Model):
+    action_date = models.DateField()
+    fiscal_year = models.IntegerField()
+    type = models.TextField()
+    pulled_from = models.TextField()
+
+    recipient_location_country_name = models.TextField()
+    recipient_location_country_code = models.TextField()
+    recipient_location_state_code = models.TextField()
+    recipient_location_county_name = models.TextField()
+    recipient_location_county_code = models.TextField()
+    recipient_location_zip5 = models.TextField()
+    recipient_location_congressional_code = models.TextField()
+    recipient_location_foreign_province = models.TextField()
+
+    pop_country_name = models.TextField()
+    pop_country_code = models.TextField()
+    pop_state_code = models.TextField()
+    pop_county_name = models.TextField()
+    pop_county_code = models.TextField()
+    pop_zip5 = models.TextField()
+    pop_congressional_code = models.TextField()
+
+    awarding_agency_id = models.IntegerField()
+    funding_agency_id = models.IntegerField()
+    awarding_toptier_agency_name = models.TextField()
+    funding_toptier_agency_name = models.TextField()
+    awarding_subtier_agency_name = models.TextField()
+    funding_subtier_agency_name = models.TextField()
+    awarding_toptier_agency_abbreviation = models.TextField()
+    funding_toptier_agency_abbreviation = models.TextField()
+    awarding_subtier_agency_abbreviation = models.TextField()
+    funding_subtier_agency_abbreviation = models.TextField()
+
+    federal_action_obligation = models.DecimalField(max_digits=20, decimal_places=2,
+                                                    blank=True, null=True)
+    counts = models.IntegerField()
+
+    class Meta:
+        managed = False
+        db_table = 'summary_transaction_geo_view'
+
+
 class AwardMatview(models.Model):
     generated_unique_award_id = models.TextField()