--- a/usaspending_api/database_scripts/matview_generator/summary_transaction_geo_view.json
+++ b/usaspending_api/database_scripts/matview_generator/summary_transaction_geo_view.json
@@ -1,0 +1,182 @@
+{
+  "final_name": "summary_transaction_geo_view",
+  "matview_sql": [
+    "SELECT",
+    "  cast(date_trunc('month', \"transaction_normalized\".\"action_date\") as date) as \"action_date\",",
+    "  \"transaction_normalized\".\"fiscal_year\",",
+    "  \"transaction_normalized\".\"type\",",
+    "  \"transaction_fpds\".\"pulled_from\",",
+    "",
+    "  recipient_location.\"location_country_code\" AS \"recipient_location_country_code\",",
+    "  recipient_location.\"country_name\" AS \"recipient_location_country_name\",",
+    "  recipient_location.\"state_code\" AS \"recipient_location_state_code\",",
+    "  recipient_location.\"county_code\" AS \"recipient_location_county_code\",",
+    "  recipient_location.\"county_name\" AS \"recipient_location_county_name\",",
+    "  recipient_location.\"congressional_code\" AS \"recipient_location_congressional_code\",",
+    "  recipient_location.\"zip5\" AS \"recipient_location_zip5\",",
+    "",
+    "  place_of_performance.\"location_country_code\" AS \"pop_country_code\",",
+    "  place_of_performance.\"country_name\" AS \"pop_country_name\",",
+    "  place_of_performance.\"state_code\" AS \"pop_state_code\",",
+    "  place_of_performance.\"county_code\" AS \"pop_county_code\",",
+    "  place_of_performance.\"county_name\" AS \"pop_county_name\",",
+    "  place_of_performance.\"congressional_code\" AS \"pop_congressional_code\",",
+    "  place_of_performance.\"zip5\" AS \"pop_zip5\",",
+    "",
+    "  \"transaction_normalized\".\"awarding_agency_id\",",
+    "  \"transaction_normalized\".\"funding_agency_id\",",
+    "  TAA.\"name\" AS awarding_toptier_agency_name,",
+    "  TFA.\"name\" AS funding_toptier_agency_name,",
+    "  SAA.\"name\" AS awarding_subtier_agency_name,",
+    "  SFA.\"name\" AS funding_subtier_agency_name,",
+    "  TAA.\"abbreviation\" AS awarding_toptier_agency_abbreviation,",
+    "  TFA.\"abbreviation\" AS funding_toptier_agency_abbreviation,",
+    "  SAA.\"abbreviation\" AS awarding_subtier_agency_abbreviation,",
+    "  SFA.\"abbreviation\" AS funding_subtier_agency_abbreviation,",
+    "",
+    "  SUM(\"transaction_normalized\".\"federal_action_obligation\") AS \"federal_action_obligation\",",
+    "  count(*) AS counts",
+    "FROM",
+    "  \"transaction_normalized\"",
+    "LEFT OUTER JOIN",
+    "  \"transaction_fabs\" ON (\"transaction_normalized\".\"id\" = \"transaction_fabs\".\"transaction_id\")",
+    "LEFT OUTER JOIN",
+    "  \"transaction_fpds\" ON (\"transaction_normalized\".\"id\" = \"transaction_fpds\".\"transaction_id\")",
+    "LEFT OUTER JOIN",
+    "  \"references_cfda\" ON (\"transaction_fabs\".\"cfda_number\" = \"references_cfda\".\"program_number\")",
+    "LEFT OUTER JOIN",
+    "  \"legal_entity\" ON (\"transaction_normalized\".\"recipient_id\" = \"legal_entity\".\"legal_entity_id\")",
+    "LEFT OUTER JOIN",
+    "  \"references_location\" AS recipient_location ON (\"legal_entity\".\"location_id\" = recipient_location.\"location_id\")",
+    "LEFT OUTER JOIN",
+    "  \"references_location\" AS place_of_performance ON (\"transaction_normalized\".\"place_of_performance_id\" = place_of_performance.\"location_id\")",
+    "LEFT OUTER JOIN",
+    "  \"agency\" AS AA ON (\"transaction_normalized\".\"awarding_agency_id\" = AA.\"id\")",
+    "LEFT OUTER JOIN",
+    "  \"toptier_agency\" AS TAA ON (AA.\"toptier_agency_id\" = TAA.\"toptier_agency_id\")",
+    "LEFT OUTER JOIN",
+    "  \"subtier_agency\" AS SAA ON (AA.\"subtier_agency_id\" = SAA.\"subtier_agency_id\")",
+    "LEFT OUTER JOIN",
+    "  \"agency\" AS FA ON (\"transaction_normalized\".\"funding_agency_id\" = FA.\"id\")",
+    "LEFT OUTER JOIN",
+    "  \"toptier_agency\" AS TFA ON (FA.\"toptier_agency_id\" = TFA.\"toptier_agency_id\")",
+    "LEFT OUTER JOIN",
+    "  \"subtier_agency\" AS SFA ON (FA.\"subtier_agency_id\" = SFA.\"subtier_agency_id\")",
+    "WHERE",
+    "  \"transaction_normalized\".\"action_date\" >= '2007-10-01' AND",
+    "  \"transaction_normalized\".\"federal_action_obligation\" IS NOT NULL",
+    "GROUP BY",
+    "  cast(date_trunc('month', \"transaction_normalized\".\"action_date\") as date),",
+    "  \"transaction_normalized\".\"fiscal_year\",",
+    "  \"transaction_normalized\".\"type\",",
+    "  \"transaction_fpds\".\"pulled_from\",",
+    "",
+    "  recipient_location.\"location_country_code\",",
+    "  recipient_location.\"country_name\",",
+    "  recipient_location.\"state_code\",",
+    "  recipient_location.\"county_code\",",
+    "  recipient_location.\"county_name\",",
+    "  recipient_location.\"congressional_code\",",
+    "  recipient_location.\"zip5\",",
+    "",
+    "  place_of_performance.\"location_country_code\",",
+    "  place_of_performance.\"country_name\",",
+    "  place_of_performance.\"state_code\",",
+    "  place_of_performance.\"county_code\",",
+    "  place_of_performance.\"county_name\",",
+    "  place_of_performance.\"congressional_code\",",
+    "  place_of_performance.\"zip5\",",
+    "",
+    "  \"transaction_normalized\".\"awarding_agency_id\",",
+    "  \"transaction_normalized\".\"funding_agency_id\",",
+    "  TAA.\"name\",",
+    "  TFA.\"name\",",
+    "  SAA.\"name\",",
+    "  SFA.\"name\",",
+    "  TAA.\"abbreviation\",",
+    "  TFA.\"abbreviation\",",
+    "  SAA.\"abbreviation\",",
+    "  SFA.\"abbreviation\"",
+    "ORDER BY",
+    "  cast(date_trunc('month', \"transaction_normalized\".\"action_date\") as date) DESC"
+  ],
+  "indexes": [
+    {
+      "name": "date",
+      "columns": [{"name": "\"action_date\"", "order": "DESC NULLS LAST"}]
+    }, {
+      "name": "fy",
+      "columns": [{"name": "\"fiscal_year\"", "order": "DESC NULLS LAST"}]
+    }, {
+      "name": "fy_type",
+      "columns": [
+          {"name": "\"fiscal_year\"", "order": "DESC NULLS LAST"},
+          {"name": "\"type\""}
+      ]
+    }, {
+      "name": "type",
+      "where":"\"type\" IS NOT NULL",
+      "columns": [{"name": "\"type\""}]
+    }, {
+      "name": "pulled_from",
+      "where": "\"pulled_from\" IS NOT NULL",
+      "columns": [{"name": "\"pulled_from\"", "order": "DESC NULLS LAST"}]
+    }, {
+      "name": "recipient_country_code",
+      "where": "\"recipient_location_country_code\" IS NOT NULL",
+      "columns": [{"name": "\"recipient_location_country_code\""}]
+    }, {
+      "name": "recipient_state_code",
+      "where": "\"recipient_location_state_code\" IS NOT NULL",
+      "columns": [{"name": "\"recipient_location_state_code\""}]
+    }, {
+      "name": "recipient_county_code",
+      "where": "\"recipient_location_county_code\" IS NOT NULL",
+      "columns": [{"name": "\"recipient_location_county_code\""}]
+    }, {
+      "name": "recipient_zip",
+      "where": "\"recipient_location_zip5\" IS NOT NULL",
+      "columns": [{"name": "\"recipient_location_zip5\""}]
+    }, {
+      "name": "pop_country_code",
+      "where": "\"pop_country_code\" IS NOT NULL",
+      "columns": [{"name": "\"pop_country_code\""}]
+    }, {
+      "name": "pop_state_code",
+      "where": "\"pop_state_code\" IS NOT NULL",
+      "columns": [{"name": "\"pop_state_code\""}]
+    }, {
+      "name": "pop_county_code",
+      "where": "\"pop_county_code\" IS NOT NULL",
+      "columns": [{"name": "\"pop_county_code\""}]
+    }, {
+      "name": "pop_zip",
+      "where": "\"pop_zip5\" IS NOT NULL",
+      "columns": [{"name": "\"pop_zip5\""}]
+    }, {
+      "name": "awarding_agency_id",
+      "where": "\"awarding_agency_id\" IS NOT NULL",
+      "columns": [{"name": "\"awarding_agency_id\"", "order": "ASC NULLS LAST"}]
+    }, {
+      "name": "funding_agency_id",
+      "where": "\"funding_agency_id\" IS NOT NULL",
+      "columns": [{"name": "\"funding_agency_id\"", "order": "ASC NULLS LAST"}]
+    }, {
+      "name": "awarding_toptier_agency_name",
+      "where": "\"awarding_toptier_agency_name\" IS NOT NULL",
+      "columns": [{"name": "\"awarding_toptier_agency_name\""}]
+    }, {
+      "name": "awarding_subtier_agency_name",
+      "where": "\"awarding_subtier_agency_name\" IS NOT NULL",
+      "columns": [{"name": "\"awarding_subtier_agency_name\""}]
+    }, {
+      "name": "funding_toptier_agency_name",
+      "where": "\"funding_toptier_agency_name\" IS NOT NULL",
+      "columns": [{"name": "\"funding_toptier_agency_name\""}]
+    }, {
+      "name": "funding_subtier_agency_name",
+      "where": "\"funding_subtier_agency_name\" IS NOT NULL",
+      "columns": [{"name": "\"funding_subtier_agency_name\""}]
+    }
+  ]
+}