## Good for recipient location and primary place of performance searches

This new materialized view was specifically targeted to increase performance when searching for just:

- primary place of performance (scope/locations)
- recipient location (scope/locations)

... additionally with any combination of the following: `action_date`*, `agencies`, and/or `type`

### Checks

- [ ] Backend Review
- [x] Create new materialized view in Dev

## Example Performance Difference
**FY2018, FY2017, FY2016, FY2015 PoP California - Over Time**
Original: (24s, 4.9s, 4.8s)
New (5s, 3.2s, 3.3s)

In this example, the best-time isn't a large improvement, however the worse-case (cold cache) is considerably better and it should be consistently low.

 
 
 
*Only action dates which start on the first of a month and end on the last day of a month